requirements.txt
PyQt5>=5.15.0
pytesseract
mss
python-dotenv
deepl
pywin32
opencv-python
pyinstaller
psycopg2-binary>=2.9
SQLAlchemy>=2.0
bcrypt>=4.1